+++ src/features/home/components/modules/video/types.ts
@@ -1,5 +1,8 @@
 import { YouTubePlayer } from 'react-youtube'
 
+import { VideoModule } from 'features/home/types'
+import { Offer } from 'shared/offer/types'
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export declare type YouTubeEvent<T = any> = {
   data: T
@@ -12,4 +15,22 @@ export enum PlayerState {
   PLAYING = 'playing',
   PAUSED = 'paused',
   ENDED = 'ended',
+}
+
+export interface VideoPlayerProps {
+  youtubeVideoId: string
+  offer?: Offer
+  onPressSeeOffer: () => void
+  moduleId: string
+  moduleName: string
+  homeEntryId: string
+}
+
+export interface VideoModalProps extends VideoModule {
+  offers: Offer[]
+  visible: boolean
+  hideModal: () => void
+  moduleId: string
+  homeEntryId: string
+  isMultiOffer: boolean
 }
